http://ftp.ksu.edu.tw/FTP/FreeBSD/distfiles/xscreensaver-4.23.tar.gz -rw-r--r-- 1 zblaxell zblaxell 4939070 Oct 22  2005 xscreensaver-4.23.tar.gz b7d3f0b4a94d879a2450a268ae941b92ca4fc39e  xscreensaver-4.23.tar.gz

+++ driver/setuid.c
@@ -97,7 +97,7 @@ setgroups_needed_p (uid_t target_group)
   if (n < 0)
     {
       char buf [1024];
-      sprintf (buf, "%s: getgroups(%d, ...)", blurb(), sizeof(groups)-1);
+      sprintf (buf, "%s: getgroups(%ld, ...)", blurb(), (long)sizeof(groups)-1);
       perror (buf);
       return 1;
     }
